+++ src/main/java/com/researchspace/api/clientmodel/IdentifiableNameable.java
@@ -1,14 +1,17 @@
 package com.researchspace.api.clientmodel;
 
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import lombok.ToString;
+import lombok.*;
+import lombok.experimental.SuperBuilder;
+
+import java.util.List;
 
 @Data
-@EqualsAndHashCode(callSuper = false)
-@ToString(callSuper = true)
+@EqualsAndHashCode(callSuper = true)
+@NoArgsConstructor
+@AllArgsConstructor
+@SuperBuilder
 public abstract class IdentifiableNameable extends Linkable {
-	private Long id = null;
-	private String globalId = null;
-	private String name = null;
+    private Long id;
+    private String globalId;
+    private String name;
 }