+++ src/main/java/com/researchspace/api/clientmodel/Linkable.java
@@ -5,9 +5,11 @@ import java.util.List;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
 
 @Data
 @NoArgsConstructor
+@SuperBuilder
 public abstract class Linkable {