+++ src/main/java/com/researchspace/api/clientmodel/inventory/Barcode.java
@@ -1,0 +1,19 @@
+package com.researchspace.api.clientmodel.inventory;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class Barcode {
+    private Long id;
+    private String data;
+    private String format;
+    private String description;
+    @Builder.Default
+    private boolean newBarcodeRequest = true;
+}